requirements.txt
BeautifulSoup==3.2.1
Django==1.8.2
beautifulsoup4==4.3.2
dj-database-url==0.3.0
dj-static==0.0.6
django-nose==1.3
django-toolbelt==0.0.1
elasticsearch==1.6.0
gunicorn==19.3.0
html5lib==0.999
jsonfield==1.0.3
lxml==3.4.4
pandas==0.16.2
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
six==1.9.0
static3==0.6.1
Unidecode==0.4.18
networkx==1.9.1
fuzzywuzzy==0.5.0
python-Levenshtein==0.12.0
djangorestframework==3.1.3
django-cors-headers==1.1.0
